+++ app/[alias]/(dashboard)/treasury/actions.tsx
@@ -1,7 +1,7 @@
 'use server';
 
-import { getServiceRoleClient } from '@/services/db';
-import { getTreasuryTransfersOfToken } from '@/services/db/transfers';
+import { getServiceRoleClient } from '@/services/chain-db';
+import { getTreasuryTransfersOfToken } from '@/services/chain-db/transfers';
 import { getAuthUserRoleInCommunityAction } from '@/app/_actions/admin-actions';
 import { Config } from '@citizenwallet/sdk';
